server/python-flask-server-generated-v03/requirements.txt
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
pyjwt == 2.3.0
cryptography >= 35.0.0
